backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
httpx==0.26.0
backoff==2.2.1
cachetools==5.3.2
slowapi==0.1.9
pydantic
python-dotenv==1.0.0
pytest==7.4.4
pytest-asyncio==0.23.3
pydantic[email]
supabase